+++ wrapper/src/main/java/software/amazon/jdbc/plugin/failover/FailoverConnectionPlugin.java
@@ -30,7 +30,6 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import software.amazon.jdbc.AwsWrapperProperty;
-import software.amazon.jdbc.HostAvailability;
 import software.amazon.jdbc.HostListProviderService;
 import software.amazon.jdbc.HostRole;
 import software.amazon.jdbc.HostSpec;
@@ -40,6 +39,7 @@ import software.amazon.jdbc.OldConnectionSuggestedAction;
 import software.amazon.jdbc.PluginManagerService;
 import software.amazon.jdbc.PluginService;
 import software.amazon.jdbc.PropertyDefinition;
+import software.amazon.jdbc.hostavailability.HostAvailability;
 import software.amazon.jdbc.plugin.AbstractConnectionPlugin;
 import software.amazon.jdbc.plugin.staledns.AuroraStaleDnsHelper;
 import software.amazon.jdbc.util.Messages;
@@ -581,7 +581,7 @@ public class FailoverConnectionPlugin extends AbstractConnectionPlugin {
 
     HostSpec failedHost = null;
     final Set<String> oldAliases = this.pluginService.getCurrentHostSpec().getAliases();
-    if (failedHostSpec != null && failedHostSpec.getAvailability() == HostAvailability.AVAILABLE) {
+    if (failedHostSpec != null && failedHostSpec.getRawAvailability() == HostAvailability.AVAILABLE) {
       failedHost = failedHostSpec;
     }
     final ReaderFailoverResult result = readerFailoverHandler.failover(this.pluginService.getHosts(), failedHost);